refactor googletest integration
