--- Src/libCZI_UnitTests/CMakeLists.txt.in
@@ -1,19 +1,0 @@
-# SPDX-FileCopyrightText: 2017-2022 Carl Zeiss Microscopy GmbH
-#
-# SPDX-License-Identifier: LGPL-3.0-or-later
-
-cmake_minimum_required(VERSION 2.8.2)
-
-project(googletest-download NONE)
-
-include(ExternalProject)
-ExternalProject_Add(googletest
-  GIT_REPOSITORY    https://github.com/google/googletest.git
-  GIT_TAG           v1.14.0
-  SOURCE_DIR        "${CMAKE_CURRENT_BINARY_DIR}/googletest-src"
-  BINARY_DIR        "${CMAKE_CURRENT_BINARY_DIR}/googletest-build"
-  CONFIGURE_COMMAND ""
-  BUILD_COMMAND     ""
-  INSTALL_COMMAND   ""
-  TEST_COMMAND      ""
-)
