--- Src/libCZI_UnitTests/pch.cpp
@@ -1,5 +1,0 @@
-// SPDX-FileCopyrightText: 2017-2022 Carl Zeiss Microscopy GmbH
-//
-// SPDX-License-Identifier: LGPL-3.0-or-later
-
-#include "pch.h"
