+++ Src/libCZI_UnitTests/testImage1.cpp
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: LGPL-3.0-or-later
 
-#include "pch.h"
+#include "include_gtest.h"
 #include "testImage.h"
 
 /*static*/ const void* CTestImage::GetZStd1CompressedImage(size_t* size, int* width, int* height)
