+++ Src/libCZI_UnitTests/test_Accessors.cpp
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: LGPL-3.0-or-later
 
-#include "pch.h"
+#include "include_gtest.h"
 #include <array>
 #include <tuple>
 #include <memory>
