+++ Src/libCZI_UnitTests/test_ZstdCompress.cpp
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: LGPL-3.0-or-later
 
-#include "pch.h"
+#include "include_gtest.h"
 #include "inc_libCZI.h"
 #include "utils.h"
 #include "../libCZI/decoder_zstd.h"
